Create default linting rules
This extracts the current set of rules from the Nutshell web app,
allowing reuse across different repos.  Rules are split into modules
based on their categorization on the [ESLint rules page](http://eslint.org/docs/rules/).

Currently only a default config is being exported, but it is possible
in the future to export different sets of rules tailored to target apps or
environments.